# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Landevo Real Estate Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /verifications:
    get:
      operationId: listVerifications
      tags: [verifications]
      summary: List verification requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Verifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Verification"
  /verifications/{verificationId}:
    get:
      operationId: getVerification
      tags: [verifications]
      summary: Get a verification by ID
      parameters:
        - in: path
          name: verificationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"
  /verifications/{verificationId}/assign:
    patch:
      operationId: assignVerification
      tags: [verifications]
      summary: Assign a verification to an officer
      parameters:
        - in: path
          name: verificationId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationAssignment"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"
  /verifications/{verificationId}/approve:
    patch:
      operationId: approveVerification
      tags: [verifications]
      summary: Approve a verification
      parameters:
        - in: path
          name: verificationId
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationReview"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"
  /verifications/{verificationId}/reject:
    patch:
      operationId: rejectVerification
      tags: [verifications]
      summary: Reject a verification
      parameters:
        - in: path
          name: verificationId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationReview"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"

  # ── Transactions ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Verification:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        listingTitle:
          type: string
        agentName:
          type: string
        officerId:
          type: ["integer", "null"]
        officerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_review, approved, rejected]
        notes:
          type: ["string", "null"]
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, listingId, listingTitle, agentName, status, submittedAt, createdAt]
    VerificationAssignment:
      type: object
      properties:
        officerId:
          type: integer
      required: [officerId]
    VerificationReview:
      type: object
      properties:
        notes:
          type: string

    # ── Transactions ──────────────────────────────────────────────────────
...